+++ lib/data/mock_job.dart
@@ -8,6 +8,7 @@ final mockJob = [
     jobType: "FullTime",
     value: 15000.0,
     imageUrl: "https://cdn-icons-png.flaticon.com/512/174/174857.png",
+    seniority: 'JR',
     modality: "Remoto",
     // Falta Regime (CLT ou PJ)
     // Informações da empresa
@@ -22,6 +23,7 @@ final mockJob = [
     value: 15000.0,
     imageUrl:
         "https://static-images.ifood.com.br/image/upload/t_high/logosgde/15def965-6971-41c7-9aa1-25f71d40f45f_CAFET_SASCO.png",
+    seniority: 'JR',
     modality: "Remoto",
   ),
   JobModel(
@@ -31,6 +33,7 @@ final mockJob = [
     jobType: "FullTime",
     value: 15000.0,
     imageUrl: "https://cdn-icons-png.flaticon.com/512/2111/2111370.png",
+    seniority: 'JR',
     modality: "Remoto",
   ),
   JobModel(
@@ -41,6 +44,7 @@ final mockJob = [
     value: 15000.0,
     imageUrl:
         "https://seeklogo.com/images/C/Coca-Cola-logo-D3CE7D7110-seeklogo.com.png",
+    seniority: 'JR',
     modality: "Remoto",
   ),
   JobModel(
@@ -50,6 +54,7 @@ final mockJob = [
     jobType: "FullTime",
     value: 15000.0,
     imageUrl: "https://cdn-icons-png.flaticon.com/512/174/174857.png",
+    seniority: 'JR',
     modality: "Remoto",
   ),
   JobModel(
@@ -60,6 +65,7 @@ final mockJob = [
     value: 15000.0,
     imageUrl:
         "https://static-images.ifood.com.br/image/upload/t_high/logosgde/15def965-6971-41c7-9aa1-25f71d40f45f_CAFET_SASCO.png",
+    seniority: 'JR',
     modality: "Remoto",
   ),
   JobModel(
@@ -69,6 +75,7 @@ final mockJob = [
     jobType: "FullTime",
     value: 15000.0,
     imageUrl: "https://cdn-icons-png.flaticon.com/512/2111/2111370.png",
+    seniority: 'JR',
     modality: "Remoto",
   ),
   JobModel(
@@ -79,6 +86,7 @@ final mockJob = [
     value: 15000.0,
     imageUrl:
         "https://seeklogo.com/images/C/Coca-Cola-logo-D3CE7D7110-seeklogo.com.png",
+    seniority: 'JR',
     modality: "Remoto",
   ),
   JobModel(
@@ -88,6 +96,7 @@ final mockJob = [
     jobType: "FullTime",
     value: 15000.0,
     imageUrl: "https://cdn-icons-png.flaticon.com/512/174/174857.png",
+    seniority: 'JR',
     modality: "Remoto",
   ),
   JobModel(
@@ -98,6 +107,7 @@ final mockJob = [
     value: 15000.0,
     imageUrl:
         "https://static-images.ifood.com.br/image/upload/t_high/logosgde/15def965-6971-41c7-9aa1-25f71d40f45f_CAFET_SASCO.png",
+    seniority: 'JR',
     modality: "Remoto",
   ),
   JobModel(
@@ -107,6 +117,7 @@ final mockJob = [
     jobType: "FullTime",
     value: 15000.0,
     imageUrl: "https://cdn-icons-png.flaticon.com/512/2111/2111370.png",
+    seniority: 'JR',
     modality: "Remoto",
   ),
   JobModel(
@@ -117,6 +128,7 @@ final mockJob = [
     value: 15000.0,
     imageUrl:
         "https://seeklogo.com/images/C/Coca-Cola-logo-D3CE7D7110-seeklogo.com.png",
+    seniority: 'JR',
     modality: "Remoto",
   ),
   JobModel(
@@ -126,6 +138,7 @@ final mockJob = [
     jobType: "FullTime",
     value: 15000.0,
     imageUrl: "https://cdn-icons-png.flaticon.com/512/174/174857.png",
+    seniority: 'JR',
     modality: "Remoto",
   ),
   JobModel(
@@ -136,6 +149,7 @@ final mockJob = [
     value: 15000.0,
     imageUrl:
         "https://static-images.ifood.com.br/image/upload/t_high/logosgde/15def965-6971-41c7-9aa1-25f71d40f45f_CAFET_SASCO.png",
+    seniority: 'JR',
     modality: "Remoto",
   ),
   JobModel(
@@ -145,6 +159,7 @@ final mockJob = [
     jobType: "FullTime",
     value: 15000.0,
     imageUrl: "https://cdn-icons-png.flaticon.com/512/2111/2111370.png",
+    seniority: 'JR',
     modality: "Remoto",
   ),
   JobModel(
@@ -155,6 +170,7 @@ final mockJob = [
     value: 15000.0,
     imageUrl:
         "https://seeklogo.com/images/C/Coca-Cola-logo-D3CE7D7110-seeklogo.com.png",
+    seniority: 'JR',
     modality: "Remoto",
   ),
   JobModel(
@@ -164,6 +180,7 @@ final mockJob = [
     jobType: "FullTime",
     value: 15000.0,
     imageUrl: "https://cdn-icons-png.flaticon.com/512/174/174857.png",
+    seniority: 'JR',
     modality: "Remoto",
   ),
   JobModel(
@@ -174,6 +191,7 @@ final mockJob = [
     value: 15000.0,
     imageUrl:
         "https://static-images.ifood.com.br/image/upload/t_high/logosgde/15def965-6971-41c7-9aa1-25f71d40f45f_CAFET_SASCO.png",
+    seniority: 'JR',
     modality: "Remoto",
   ),
   JobModel(
@@ -183,6 +201,7 @@ final mockJob = [
     jobType: "FullTime",
     value: 15000.0,
     imageUrl: "https://cdn-icons-png.flaticon.com/512/2111/2111370.png",
+    seniority: 'JR',
     modality: "Remoto",
   ),
   JobModel(
@@ -193,6 +212,7 @@ final mockJob = [
     value: 15000.0,
     imageUrl:
         "https://seeklogo.com/images/C/Coca-Cola-logo-D3CE7D7110-seeklogo.com.png",
+    seniority: 'JR',
     modality: "Remoto",
   ),
 ];
