+++ lib/models/job_model.dart
@@ -7,6 +7,7 @@ class JobModel {
   double value;
   String imageUrl;
   String modality;
+  String seniority;
 
   JobModel({
     required this.jobOportunity,
@@ -16,5 +17,7 @@ class JobModel {
     required this.value,
     required this.imageUrl,
     required this.modality,
+    required this.seniority,
+    
   });
 }
