+++ lib/views/vacancy_details/welcome_page.dart
